server/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Mail==0.10.0
Werkzeug~=3.0.1
bcrypt==4.2.0
pyjwt==2.4.0
mysqlclient==2.2.4
Sphinx
sphinx-rtd-theme
SQLAlchemy
python-dotenv